fix : Step4 리뷰 보완점 수정
