+++ src/main/java/carracing/domain/car/strategy/CarCondition.java
@@ -1,4 +1,4 @@
-package carracing.domain;
+package carracing.domain.car.strategy;
 
 public interface CarCondition {
     boolean isMovable();
